# Optimize addToTree function.