--- a/src/utils/sources-tree/tests/utils.spec.js
+++ b/src/utils/sources-tree/tests/utils.spec.js
@@ -50,7 +50,7 @@ describe("sources tree", () => {
       ];
 
       const tree = createNode("root", "", []);
-      sources.forEach(source => addToTree(tree, source));
+      sources.forEach(source => addToTree(tree, source, "http://example.com/"));
       sortEntireTree(tree);
       const [bFolderNode, aFileNode] = tree.contents[0].contents;
       const [cFolderNode] = bFolderNode.contents;
@@ -101,9 +101,9 @@ describe("sources tree", () => {
       });
 
       const tree = createNode("root", "", []);
-      addToTree(tree, source1);
-      addToTree(tree, source2);
-      addToTree(tree, source3);
+      addToTree(tree, source1, "http://a/");
+      addToTree(tree, source2, "http://a/");
+      addToTree(tree, source3, "http://a/");
       const paths = getDirectories("http://a/b.js", tree);
 
       expect(paths[1].path).toBe("/a");
@@ -122,8 +122,8 @@ describe("sources tree", () => {
       });
 
       const tree = createNode("root", "", []);
-      addToTree(tree, source1);
-      addToTree(tree, source2);
+      addToTree(tree, source1, "http://a/");
+      addToTree(tree, source2, "http://a/");
       const paths = getDirectories("http://a/b.js?key=hi", tree);
 
       expect(paths[1].path).toBe("/a");
@@ -142,8 +142,8 @@ describe("sources tree", () => {
       });
 
       const tree = createNode("root", "", []);
-      addToTree(tree, source1);
-      addToTree(tree, source2);
+      addToTree(tree, source1, "http://a/");
+      addToTree(tree, source2, "http://a/");
       const paths = getDirectories("https://a/b.js", tree);
 
       expect(paths[1].path).toBe("/a");