… src/utils/sources-tree/tests/__snapshots__/addToTree.spec.js.snap