Associated Issue: #4168

### Summary of Changes

* don't use sortTree  for sources
* improve the performance of addToTree

### Test Plan

Added console.time() in the SourcesTree.js and createTree.js around addToTree+sortTree operations.
Measured performance of https://yurydelendik.github.io/old-man-sandbox/create-30k/f3k.html tree.

Before:
<img width="1174" alt="screen shot 2017-10-03 at 2 59 13 pm" src="https://user-images.githubusercontent.com/1523410/31150699-422e6fd4-a85a-11e7-89a6-7a150fd7c2e9.png">

After:
<img width="1181" alt="screen shot 2017-10-03 at 3 03 50 pm" src="https://user-images.githubusercontent.com/1523410/31150728-5c2d7d3a-a85a-11e7-8036-aba6f0e5225d.png">

Performance improved from 1186ms->291ms also affected subsequent/incremental tree update operations 180ms->2ms.