@codehag 2017-10-04
Thanks for the clarifications yury! i made one more comment for your feedback